refactor(acp): move the ACP proxy engine into bitrouter-sdk
The substrate was the ACP executor + proxy engine — the direct analogue of
the MCP `RmcpExecutor` that already lives inside the SDK — yet the SDK's
`acp` module pointed outward to it for the runtime. That inversion is what
this moves: `up`, `executor`, `turn`, `telemetry`, `permissions`, `session`,
`engine` (lean `Session`), and `down` (`serve`/`serve_on`) now live under
`bitrouter_sdk::acp`, behind the `acp` feature exactly as `rmcp_executor`
rides `mcp`. Without the feature the SDK still exposes `AcpTransport`,
`AcpTarget`, and the `RoutingTable` trait and pulls none of the new deps.

`bitrouter-substrate` is now a shim that re-exports each moved module by
module path, so every `bitrouter_substrate::<module>::Item` importer keeps
compiling until the next commit migrates them.

Two things could not move verbatim:

- `bitrouter-sdk` is `#![forbid(unsafe_code)]`, so `up`'s raw
  `libc::killpg` becomes `rustix::process::kill_process_group` — a safe
  wrapper over the same syscall, from a crate already in the build graph.
- The SDK warns on `missing_docs`, so the moved public items are documented.

`SessionStatus` goes with the move: nothing ever read `SessionState::status`
and only two of its six variants were ever constructed. A lean `Session`
does not carry state no one observes.
