+++ Cargo.toml
@@ -37,6 +37,10 @@ bitrouter-substrate = { path = "crates/bitrouter-substrate", version = "1.0.0-al
 bitrouter-guardrails = { path = "crates/bitrouter-guardrails", version = "1.0.0-alpha.27" }
 bitrouter-observe = { path = "crates/bitrouter-observe", version = "1.0.0-alpha.27", default-features = false }
 
+# Safe syscall wrappers. Used by the SDK's `acp` runtime for `killpg(2)`;
+# `libc` would need an `unsafe` block, and `bitrouter-sdk` forbids unsafe code.
+rustix = { version = "1", default-features = false }
+
 # ACP SDK crates
 agent-client-protocol = "1.2"
 agent-client-protocol-schema = "1.4"
