+++ crates/bitrouter-substrate/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bitrouter-substrate"
-description = "Per-session ACP substrate: one session, one agent, turn queue"
+description = "Transitional shim: the ACP thin proxy now lives in bitrouter-sdk::acp"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -10,21 +10,6 @@ rust-version.workspace = true
 
 [dependencies]
 bitrouter-sdk = { workspace = true, features = ["acp"] }
-agent-client-protocol = { workspace = true }
-agent-client-protocol-schema = { workspace = true }
-tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "macros", "process", "io-util"] }
-tokio-stream = { workspace = true, features = ["sync"] }
-tokio-util = { workspace = true, features = ["compat"] }
-futures.workspace = true
-async-trait.workspace = true
-anyhow.workspace = true
-tracing.workspace = true
-uuid.workspace = true
-serde = { workspace = true }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
 
 [dev-dependencies]
-tokio = { workspace = true, features = ["rt-multi-thread", "test-util"] }
 tempfile.workspace = true
